app/shared/ai.py
"""Vertex AI操作ライブラリ."""

from google.cloud import aiplatform
import logging
from typing import Dict, Any

logger = logging.getLogger(__name__)


class VertexAIClient:
    def __init__(self, project_id: str, location: str, model_name: str):
        self.project_id = project_id
        self.location = location
        self.model_name = model_name
        aiplatform.init(project=project_id, location=location)

    def analyze_audio(self, audio_uri: str) -> Dict[str, Any]:
        """Vertex AI を使用して音声ファイルを分析.

        注: このはサンプル実装。実際の Gemini API 呼び出しは別途実装が必要.
        """
        logger.info(f"Analyzing audio: {audio_uri} with model {self.model_name}")

        try:
            # 実装例: Gemini 1.5 Pro で音声から情報を抽出
            # 実際の実装は google.generativeai や Vertex AI API を使用

            # ここはプレースホルダ
            result = {
                "title": "Sample Podcast Title",
                "summary": "This is a sample summary of the podcast.",
                "transcript": "Sample transcript content...",
                "duration_seconds": 600,
                "keywords": ["keyword1", "keyword2"],
            }

            logger.info(f"Analysis completed: {result}")
            return result
        except Exception as e:
            logger.error(f"Failed to analyze audio: {e}")
            raise

    def process_audio_file(self, gcs_uri: str) -> Dict[str, Any]:
        """GCS 上の音声ファイルを処理."""
        return self.analyze_audio(gcs_uri)
