pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-automator"
description = "Podcast automation system using GCP Cloud Run Jobs and Cloudflare R2"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your-email@example.com"}
]
requires-python = ">=3.11"

dependencies = [
    "google-cloud-storage>=2.10",
    "google-cloud-aiplatform>=1.25",
    "google-cloud-secret-manager>=2.16",
    "google-cloud-logging>=3.5",
    "google-cloud-run>=0.8",
    "flask>=2.3",
    "gunicorn>=20.1",
    "boto3>=1.26",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pylint>=2.17",
    "flake8>=6.0",
    "hypothesis>=6.80",
]

test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]

[tool.setuptools]
packages = ["shared"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long
    "E731",  # do not assign a lambda expression, use a def
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["app"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app/shared --cov-report=html --cov-report=term-missing"
